updated frontend code on 1 Augest

+++ src/app/Components/veiw-balance/veiw-balance.component.ts
@@ -1,7 +1,10 @@
 import { HttpClient } from '@angular/common/http';
 import { Component, OnInit } from '@angular/core';
 import { FormBuilder, Validators } from '@angular/forms';
+import { Router } from '@angular/router';
 import { BalanceService } from 'src/app/balance.service';
+import { User } from 'src/app/Component/sign-in/User';
+import { NewUser } from 'src/app/user/User';
 
 @Component({
   selector: 'app-veiw-balance',
@@ -15,25 +18,38 @@ export class VeiwBalanceComponent implements OnInit {
   loginForm: any;
   submitted!: boolean;
   router: any;
-  
-  constructor(private service :BalanceService, private formBuilder: FormBuilder
+  response:any;
+  checkUser !: NewUser;
+  panNo:any;
+  user : any;
+  constructor(private service :BalanceService, private formBuilder: FormBuilder, private rout : Router
     ) { }
 
   ngOnInit(){
-    let mobile_no= localStorage.getItem("mobileNo") ; 
-    let pan_no = localStorage.getItem("panNo"); 
-    let date_of_birth = localStorage.getItem("dateOfBirth"); 
-   this.service.getBalances(mobile_no,pan_no,date_of_birth).subscribe(data=>
-      {
-       if(this.listOfBalances!=null){
-        alert("This user has not registered for any account")
-       }else
-       {
-        this.listOfBalances.push(data);
-       }
-        console.log("view comp data"+this.listOfBalances);
-      });
-        localStorage.clear();
+    let mobile_no= localStorage.getItem("mobileNo") as any ; 
+    let pan_no = localStorage.getItem("panNo") as any; 
+    let date_of_birth = localStorage.getItem("dateOfBirth") as any; 
+    let user_name=localStorage.getItem("username")  as string;
+    this.service.getUserDetails(user_name)
+          .subscribe(
+              data => {
+                this.response=data;
+                console.log("from service"+data);
+               this.user= data;
+               this.checkUser = this.user;
+               console.log("Mobile number: ."+this.checkUser.mobileNo);
+               if(this.checkUser.mobileNo==mobile_no && this.checkUser.panNo==pan_no && this.checkUser.dateOfBirth==date_of_birth){
+                this.service.getBalances(this.checkUser.mobileNo, this.checkUser.panNo,this.checkUser.dateOfBirth).subscribe(data=>
+                  {
+                      this.listOfBalances.push(data);
+                    console.log("view comp data"+this.listOfBalances);
+                  });
+                }else{
+                  alert("You have entered wrong credential")
+                }
+                    localStorage.clear();
+               
+              });
       }
           
     }
